+++ tests/Fixtures/views/custom-scripts.blade.php
@@ -1,1 +1,0 @@
-<script data-custom-model-mind-scripts></script>