+++ tests/Fixtures/views/custom-styles.blade.php
@@ -1,1 +1,0 @@
-<style data-custom-model-mind-styles></style>